DAVAI-76 Streamline versioning, fix environment vars
This removes the version from the package.json and App.tsx.
When the plugin is built the build system will get the version string from the DEPLOY_PATH which will either be a branch or tag. In the case of a real version it will be a tag.

Previously the webpack config was including every environment variable in the build environment into the built file. This risks exposing secrets from the build environment to the public.