DAVAI-84 sonify collections correctly
When a graph is showing attributes that are a collection, it only shows points for the cases in the most nested collection that has an attribute.
So if there is a more nested collection, the graph will not show duplicate points for those cases.
The sonification code now works the same way.

This also updates the way the graph info is updated. It uses a map and applySnapshot to automatically create, update, and delete the CODAPGraphModel objects.